Moved pin/flag dropDown menu into its own component
